How to Split a Python List - Update materials
