+++ python-split-list/README.md
@@ -1,3 +1,3 @@
-# How to Split a Python List
+# How to Split a Python List Into Chunks
 
-This folder holds the code for the Real Python How to Split a Python List tutorial.
+This folder holds sample code that supplements the Real Python tutorial on [How to Split a Python List Into Chunks](https://realpython.com/how-to-split-a-python-list-into-chunks/).
