+++ python-split-list/requirements.txt
@@ -1,3 +1,3 @@
-Pillow==9.3.0
+Pillow==9.4.0
 more-itertools==9.0.0
-numpy==1.23.5
+numpy==1.24.1
