+++ python-split-list/spatial_splitting.py
